requirements.txt
websockets==12.0
deepgram-sdk
requests